requirements.txt
fastapi>=0.115.11
uvicorn>=0.24.0
redis>=5.0.1
openai>=1.3.7
nanoid>=2.0.0
pydantic>=2.5.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
tiktoken>=0.5.1
structlog>=23.2.0
async-timeout>=4.0.3
httpx>=0.25.1 
numpy
